+++ server/src/main/kotlin/org/javacs/kt/KotlinLanguageServer.kt
@@ -8,8 +8,11 @@ import org.eclipse.lsp4j.services.LanguageClientAware
 import org.eclipse.lsp4j.services.LanguageServer
 import org.javacs.kt.commands.ALL_COMMANDS
 import org.javacs.kt.externalsources.JarClassContentProvider
+import org.javacs.kt.util.AsyncExecutor
 import org.javacs.kt.util.TemporaryDirectory
 import org.javacs.kt.util.parseURI
+import org.javacs.kt.progress.Progress
+import org.javacs.kt.progress.LanguageClientProgress
 import java.net.URI
 import java.io.Closeable
 import java.nio.file.Paths
@@ -17,20 +20,38 @@ import java.util.concurrent.CompletableFuture
 import java.util.concurrent.CompletableFuture.completedFuture
 
 class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
-    private val config = Configuration()
+    val config = Configuration()
     val classPath = CompilerClassPath(config.compiler)
 
     private val tempDirectory = TemporaryDirectory()
     private val uriContentProvider = URIContentProvider(JarClassContentProvider(config.externalSources, classPath, tempDirectory))
-    val sourcePath = SourcePath(classPath, uriContentProvider)
+    val sourcePath = SourcePath(classPath, uriContentProvider, config.indexing)
     val sourceFiles = SourceFiles(sourcePath, uriContentProvider)
 
     private val textDocuments = KotlinTextDocumentService(sourceFiles, sourcePath, config, tempDirectory, uriContentProvider)
     private val workspaces = KotlinWorkspaceService(sourceFiles, sourcePath, classPath, textDocuments, config)
     private val protocolExtensions = KotlinProtocolExtensionService(uriContentProvider)
 
+    private lateinit var client: LanguageClient
+
+    private val async = AsyncExecutor()
+    private var progressFactory: Progress.Factory = Progress.Factory.None
+        set(factory: Progress.Factory) {
+            field = factory
+            sourcePath.progressFactory = factory
+        }
+
+    companion object {
+        val VERSION: String? = System.getProperty("kotlinLanguageServer.version")
+    }
+
+    init {
+        LOG.info("Kotlin Language Server: Version ${VERSION ?: "?"}")
+    }
+
     override fun connect(client: LanguageClient) {
-        connectLoggingBackend(client)
+        this.client = client
+        connectLoggingBackend()
 
         workspaces.connect(client)
         textDocuments.connect(client)
@@ -45,7 +66,7 @@ class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
     @JsonDelegate
     fun getProtocolExtensionService(): KotlinProtocolExtensions = protocolExtensions
 
-    override fun initialize(params: InitializeParams): CompletableFuture<InitializeResult> {
+    override fun initialize(params: InitializeParams): CompletableFuture<InitializeResult> = async.compute {
         val serverCapabilities = ServerCapabilities()
         serverCapabilities.setTextDocumentSync(TextDocumentSyncKind.Incremental)
         serverCapabilities.workspace = WorkspaceServerCapabilities()
@@ -67,22 +88,39 @@ class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
         val clientCapabilities = params.capabilities
         config.completion.snippets.enabled = clientCapabilities?.textDocument?.completion?.completionItem?.snippetSupport ?: false
 
-        if (params.rootUri != null) {
-            LOG.info("Adding workspace {} to source path", params.rootUri)
+        if (clientCapabilities?.window?.workDoneProgress ?: false) {
+            progressFactory = LanguageClientProgress.Factory(client)
+        }
 
-            val root = Paths.get(parseURI(params.rootUri))
+        val folders = params.workspaceFolders?.takeIf { it.isNotEmpty() }
+            ?: params.rootUri?.let(::WorkspaceFolder)?.let(::listOf)
+            ?: listOf()
 
+        val progress = params.workDoneToken?.let { LanguageClientProgress("Workspace folders", it, client) }
+
+        folders.forEachIndexed { i, folder ->
+            LOG.info("Adding workspace folder {}", folder.name)
+            val progressPrefix = "[${i + 1}/${folders.size}] ${folder.name}"
+            val progressPercent = (100 * i) / folders.size
+
+            progress?.update("$progressPrefix: Updating source path", progressPercent)
+            val root = Paths.get(parseURI(folder.uri))
             sourceFiles.addWorkspaceRoot(root)
+
+            progress?.update("$progressPrefix: Updating class path", progressPercent)
             val refreshed = classPath.addWorkspaceRoot(root)
             if (refreshed) {
+                progress?.update("$progressPrefix: Refreshing source path", progressPercent)
                 sourcePath.refresh()
             }
         }
 
-        return completedFuture(InitializeResult(serverCapabilities))
+        val serverInfo = ServerInfo("Kotlin Language Server", VERSION)
+
+        InitializeResult(serverCapabilities, serverInfo)
     }
 
-    private fun connectLoggingBackend(client: LanguageClient) {
+    private fun connectLoggingBackend() {
         val backend: (LogMessage) -> Unit = {
             client.logMessage(MessageParams().apply {
                 type = it.level.toLSPMessageType()
@@ -104,6 +142,7 @@ class KotlinLanguageServer : LanguageServer, LanguageClientAware, Closeable {
         textDocumentService.close()
         classPath.close()
         tempDirectory.close()
+        async.shutdown(awaitTermination = true)
     }
 
     override fun shutdown(): CompletableFuture<Any> {