+++ shared/src/main/kotlin/org/javacs/kt/classpath/DefaultClassPathResolver.kt
@@ -6,7 +6,6 @@ import org.javacs.kt.getSettingParam
 import org.javacs.kt.SettingParam
 import java.nio.file.Path
 import java.nio.file.PathMatcher
-import java.nio.file.Files
 import java.nio.file.FileSystems
 
 fun defaultClassPathResolver(workspaceRoots: Collection<Path>): ClassPathResolver =
@@ -17,7 +16,7 @@ fun defaultClassPathResolver(workspaceRoots: Collection<Path>): ClassPathResolve
 
 /** Searches the workspace for all files that could provide classpath info. */
 private fun workspaceResolvers(workspaceRoot: Path): Sequence<ClassPathResolver> {
-    val ignored: List<PathMatcher> = ignoredPathPatterns(workspaceRoot.resolve(".gitignore"))
+    val ignored: List<PathMatcher> = ignoredPathPatterns(workspaceRoot, workspaceRoot.resolve(".gitignore"))
     return folderResolvers(workspaceRoot, ignored).asSequence()
 }
 
@@ -28,34 +27,40 @@ private fun folderResolvers(root: Path, ignored: List<PathMatcher>): Collection<
 
     return root.toFile()
         .walk()
+<<<<<<< HEAD
         .onEnter { file -> ignored.none { it.matches(root.relativize(file.toPath())) } }
         .mapNotNull { file ->
             if (!whitelistOnly || whitelist.any { file.toPath().toAbsolutePath().toString().contains(it) }) {
                 asClassPathProvider(file.toPath())
             } else null
         }
+=======
+        .onEnter { file -> ignored.none { it.matches(file.toPath()) } }
+        .mapNotNull { asClassPathProvider(it.toPath()) }
+>>>>>>> 8cc94b996b600a70b65d79b067e9ea750fdbd9d5
         .toList()
     }
 
 /** Tries to read glob patterns from a gitignore. */
-private fun ignoredPathPatterns(path: Path): List<PathMatcher> =
-    path.toFile()
+private fun ignoredPathPatterns(root: Path, gitignore: Path): List<PathMatcher> =
+    gitignore.toFile()
         .takeIf { it.exists() }
         ?.readLines()
         ?.map { it.trim() }
         ?.filter { it.isNotEmpty() && !it.startsWith("#") }
+        ?.map { it.removeSuffix("/") }
         ?.let { it + listOf(
             // Patterns that are ignored by default
             ".git"
         ) }
         ?.mapNotNull { try {
-            LOG.debug("Adding ignore pattern '{}' from {}", it, path)
-            FileSystems.getDefault().getPathMatcher("glob:$it")
+            LOG.debug("Adding ignore pattern '{}' from {}", it, gitignore)
+            FileSystems.getDefault().getPathMatcher("glob:$root**/$it")
         } catch (e: Exception) {
             LOG.warn("Did not recognize gitignore pattern: '{}' ({})", it, e.message)
             null
         } }
-        ?: emptyList<PathMatcher>()
+        ?: emptyList()
 
 /** Tries to create a classpath resolver from a file using as many sources as possible */
 private fun asClassPathProvider(path: Path): ClassPathResolver? =